Finish function dropdown list
